requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
aiohttp
python-multipart
openai
langchain
langchain_community
langchain-openai
langchain-core
motor